requirements.txt
appwrite==6.1.0
fastapi
pydantic
uvicorn
python-multipart
python-dotenv
aiofiles
httpx
jinja2
rich
email-validator
pydantic_settings

